Move tenancy/utils to db/tenancy to resolve circular imports between tenancy and db package. This should remove all reliance on the tenancy package from the db package
