+++ packages/backend-core/src/tenancy/index.ts
@@ -1,11 +1,9 @@
 import * as context from "../context"
 import * as tenancy from "./tenancy"
-import * as utils from "./utils"
 
 const pkg = {
   ...context,
   ...tenancy,
-  ...utils,
 }
 
 export = pkg
