+++ packages/server/src/api/controllers/cloud.js
@@ -1,15 +1,12 @@
 const env = require("../../environment")
-const { getAllApps } = require("@budibase/backend-core/db")
+const { getAllApps, getGlobalDBName } = require("@budibase/backend-core/db")
 const {
   exportDB,
   sendTempFile,
   readFileSync,
 } = require("../../utilities/fileSystem")
 const { stringToReadStream } = require("../../utilities")
-const {
-  getGlobalDBName,
-  getGlobalDB,
-} = require("@budibase/backend-core/tenancy")
+const { getGlobalDB } = require("@budibase/backend-core/tenancy")
 const { create } = require("./application")
 const { getDocParams, DocumentType, isDevAppID } = require("../../db/utils")
